refactor: rename config folder into legacy
