+++ packages/config/legacy/scaffold.js
