+++ packages/config/test/index.spec.js
@@ -8,8 +8,8 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const Config = require("../src/Config");
-const Settings = require("../src/Settings");
+const Config = require("../legacy/Config");
+const Settings = require("../legacy/Settings");
 const index = require("../index");
 
 describe("index", () => {
