+++ packages/config/test/legacy/CommandLineArguments.mocks.js
@@ -1,5 +1,6 @@
 /*
 Copyright 2019 Javier Brea
+Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -10,26 +11,26 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 const sinon = require("sinon");
 
-jest.mock("../../src/Config");
+jest.mock("../../legacy/CommandLineArguments");
 
-const Config = require("../../src/Config");
+const CommandLineArguments = require("../../legacy/CommandLineArguments");
 
-class ConfigMock {
+class Mock {
   constructor() {
     this._sandbox = sinon.createSandbox();
 
     this._stubs = {
-      init: this._sandbox.stub().resolves(),
-      coreOptions: {},
+      init: this._sandbox.stub(),
+      addCustom: this._sandbox.stub(),
       options: {},
     };
 
-    Config.mockImplementation(() => this._stubs);
+    CommandLineArguments.mockImplementation(() => this._stubs);
   }
 
   get stubs() {
     return {
-      Constructor: Config,
+      Constructor: CommandLineArguments,
       instance: this._stubs,
     };
   }
@@ -39,4 +40,4 @@ class ConfigMock {
   }
 }
 
-module.exports = ConfigMock;
+module.exports = Mock;
