+++ packages/config/test/legacy/config/config.add-plugins.js
