+++ packages/config/test/legacy/config/config.export-error.js
