+++ packages/config/test/legacy/config/config.export-invalid.js
