+++ packages/config/test/legacy/config/config.function-error.js
