+++ packages/config/test/legacy/config/config.function-invalid.js
