+++ packages/config/test/legacy/config/config.function.js
