+++ packages/config/test/legacy/config/config.object.js
