+++ packages/config/test/legacy/config/config.promise-invalid.js
