+++ packages/config/test/legacy/config/config.promise-rejected.js
