+++ packages/config/test/legacy/config/config.promise.js
